Update docs, include loop/trap, add sleep
The script now has proper doxygen-style documentation.  Woo hoo.

The main loop was refactored to allow trapping on SIGINT (and others) so
that one can send a signal to the script to make it stop and not
restart.

Should flask croak, sleep for a few seconds before trying to restart it.