Factor in retry/delay on failed readings
Previously, if the sensor returned None for temperature or humidity,
the program would just return an empty JSON object... which was fine when
looking at values in aggregate, but a little irritating when trying to
test and/or debug (i.e., is the program working?  is the sensor working?
where should I start looking?).

So, now if either value is `None` then sleep a second and try again.